Fix UserContext.breakLoop, update deterministic scheduling, reorganize worklist specializations
* Break loop now works
* Fixed some build bugs in sched.cpp and unsigned warnings
* Clean up some includes in Galois.h
* Added UserContextAccess to simplify friend access for UserContext
* Fixed bug in deterministic scheduler and add parallel sort
* Move inlined ForEachWork into exp


git-svn-id: https://svn.ices.utexas.edu/repos/iss/projects/GaloisCpp/trunk@9871 522c16a8-e32f-0410-b4cc-8071c0b41197

+++ include/Galois/Runtime/WorkList.h
@@ -374,7 +374,7 @@ class OrderedByIntegerMetric : private boost::noncopyable {
 #else
     unsigned msS = p.scanStart;
     if (localLeader)
-      for (int i = 0; i < ThreadPool::getActiveThreads(); ++i)
+      for (int i = 0; i < (int) ThreadPool::getActiveThreads(); ++i)
 	msS = std::min(msS, current.getRemote(i)->scanStart);
     else
       msS = std::min(msS, current.getRemote(LL::getLeaderForPackage(myID))->scanStart);
@@ -575,7 +575,7 @@ public:
 
   template<typename Iter>
   void push_initial(Iter b, Iter e) {
-    fill_work(*this,b,e);
+    fill_work(*this, b, e);
   }
 
   boost::optional<value_type> pop()  {
@@ -847,7 +847,7 @@ public:
   template<typename ItTy>
   void push_initial(ItTy b, ItTy e) {
     fill_work(*this,b,e);
-    for (int x = 0; x < pushBuffer.size(); ++x)
+    for (unsigned int x = 0; x < pushBuffer.size(); ++x)
       pushBuffer.get(x).flush();
   }
